Support JDK's LongAdder
Add an ability to use the JDK's implementation if it's available in runtime.
Add a factory for creating LongAdders which tries to use the JDK's implementation
and fallbacks to the internal one if the JDK doesn't provide any. Unfortunatelly,
LongAdders don't have a common interface, therefore we introduce `LongAdderAdapter`
as a common interface for LongAdders. It will be used in other components as interface,
and a concrete implementation of it will be created by `LongAdderFactory`.